+++ Src/MoneyFox.Ui.Tests/ViewModels/Budget/EditBudgetViewModelTests.cs
@@ -13,6 +13,7 @@ using FluentAssertions;
 using MediatR;
 using NSubstitute;
 using Views.Budget;
+using Views.Budget.ModifyBudget;
 using Views.Categories;
 using Views.Categories.CategorySelection;
 using Xunit;
