+++ Src/MoneyFox.Ui/InversionOfControl/MoneyFoxConfig.cs
@@ -11,6 +11,7 @@ using Views.About;
 using Views.Accounts;
 using Views.Backup;
 using Views.Budget;
+using Views.Budget.ModifyBudget;
 using Views.Categories;
 using Views.Categories.CategorySelection;
 using Views.Categories.ModifyCategory;
