+++ Src/MoneyFox.Ui/MoneyFox.Ui.csproj
@@ -145,6 +145,18 @@
           <DependentUpon>ModifyPaymentContentView.xaml</DependentUpon>
           <SubType>Code</SubType>
         </Compile>
+        <Compile Update="Views\Budget\ModifyBudget\ModifyBudgetView.xaml.cs">
+          <DependentUpon>ModifyBudgetView.xaml</DependentUpon>
+          <SubType>Code</SubType>
+        </Compile>
+        <Compile Update="Views\Budget\ModifyBudget\EditBudgetPage.xaml.cs">
+          <DependentUpon>EditBudgetPage.xaml</DependentUpon>
+          <SubType>Code</SubType>
+        </Compile>
+        <Compile Update="Views\Budget\ModifyBudget\AddBudgetPage.xaml.cs">
+          <DependentUpon>AddBudgetPage.xaml</DependentUpon>
+          <SubType>Code</SubType>
+        </Compile>
     </ItemGroup>
 
     <ItemGroup>
@@ -170,6 +182,15 @@
           <SubType>Designer</SubType>
         </MauiXaml>
         <MauiXaml Update="Views\Payments\PaymentModification\ModifyPaymentContentView.xaml">
+          <SubType>Designer</SubType>
+        </MauiXaml>
+        <MauiXaml Update="Views\Budget\ModifyBudget\ModifyBudgetView.xaml">
+          <SubType>Designer</SubType>
+        </MauiXaml>
+        <MauiXaml Update="Views\Budget\ModifyBudget\EditBudgetPage.xaml">
+          <SubType>Designer</SubType>
+        </MauiXaml>
+        <MauiXaml Update="Views\Budget\ModifyBudget\AddBudgetPage.xaml">
           <SubType>Designer</SubType>
         </MauiXaml>
     </ItemGroup>
