+++ Src/MoneyFox.Ui/Routes.cs
@@ -4,6 +4,7 @@ using Views.About;
 using Views.Accounts;
 using Views.Backup;
 using Views.Budget;
+using Views.Budget.ModifyBudget;
 using Views.Categories;
 using Views.Categories.CategorySelection;
 using Views.Categories.ModifyCategory;
