+++ Src/MoneyFox.Ui/Views/Budget/ModifyBudget/AddBudgetPage.xaml
@@ -4,13 +4,13 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
-    xmlns:budget="clr-namespace:MoneyFox.Ui.Views.Budget"
+    xmlns:modifyBudget="clr-namespace:MoneyFox.Ui.Views.Budget.ModifyBudget"
     Title="{x:Static resources:Translations.AddBudgetTitle}"
-    x:DataType="budget:AddBudgetViewModel"
-    x:Class="MoneyFox.Ui.Views.Budget.AddBudgetPage">
+    x:DataType="modifyBudget:AddBudgetViewModel"
+    x:Class="MoneyFox.Ui.Views.Budget.ModifyBudget.AddBudgetPage">
 
     <Grid RowDefinitions="*, 6, 40" Padding="12">
-        <budget:ModifyBudgetView Grid.Row="0" />
+        <modifyBudget:ModifyBudgetView Grid.Row="0" />
         <Button Grid.Row="2"
                 Text="{x:Static resources:Translations.SaveLabel}"
                 Command="{Binding SaveBudgetCommand}"
