+++ Src/MoneyFox.Ui/Views/Budget/ModifyBudget/BudgetTimeRangeStringConverter.cs
@@ -1,8 +1,8 @@
-namespace MoneyFox.Ui.Views.Budget;
+namespace MoneyFox.Ui.Views.Budget.ModifyBudget;
 
 using System.Globalization;
-using Domain.Aggregates.BudgetAggregate;
-using Resources.Strings;
+using MoneyFox.Domain.Aggregates.BudgetAggregate;
+using MoneyFox.Ui.Resources.Strings;
 
 internal class BudgetTimeRangeStringConverter : IValueConverter
 {
