+++ Src/MoneyFox.Ui/Views/Budget/ModifyBudget/ModifyBudgetView.xaml
@@ -6,14 +6,14 @@
     xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
     xmlns:converter="clr-namespace:MoneyFox.Ui.Converter"
     xmlns:controls="clr-namespace:MoneyFox.Ui.Controls"
-    xmlns:budget="clr-namespace:MoneyFox.Ui.Views.Budget"
-    x:DataType="budget:ModifyBudgetViewModel"
-    x:Class="MoneyFox.Ui.Views.Budget.ModifyBudgetView">
+    xmlns:modifyBudget="clr-namespace:MoneyFox.Ui.Views.Budget.ModifyBudget"
+    x:DataType="modifyBudget:ModifyBudgetViewModel"
+    x:Class="MoneyFox.Ui.Views.Budget.ModifyBudget.ModifyBudgetView">
 
     <ContentView.Resources>
         <ResourceDictionary>
             <converter:DecimalConverter x:Key="DecimalConverter" />
-            <budget:BudgetTimeRangeStringConverter x:Key="BudgetTimeRangeStringConverter" />
+            <modifyBudget:BudgetTimeRangeStringConverter x:Key="BudgetTimeRangeStringConverter" />
         </ResourceDictionary>
     </ContentView.Resources>
 
@@ -33,7 +33,7 @@
                Text="{x:Static resources:Translations.EvaluationTimeRangeLabel}" />
 
         <Picker Title="{x:Static resources:Translations.EvaluationTimeRangeLabel}"
-                ItemsSource="{x:Static budget:ModifyBudgetViewModel.TimeRangeCollection}"
+                ItemsSource="{x:Static modifyBudget:ModifyBudgetViewModel.TimeRangeCollection}"
                 SelectedItem="{Binding TimeRange}"
                 ItemDisplayBinding="{Binding ., Converter={StaticResource BudgetTimeRangeStringConverter}}" />
 
@@ -51,7 +51,7 @@
                         SelectionChangedCommand="{Binding RemoveCategoryCommand}"
                         SelectionChangedCommandParameter="{Binding SelectedItem, Source={x:Reference BudgetCategoryCollectionView}}">
             <CollectionView.ItemTemplate>
-                <DataTemplate x:DataType="budget:BudgetCategoryViewModel">
+                <DataTemplate x:DataType="modifyBudget:BudgetCategoryViewModel">
                     <Grid Padding="3" HeightRequest="60">
                         <controls:Card>
                             <Label Text="{Binding Name}" VerticalOptions="Center" HorizontalOptions="Center" />
